pyproject.toml
[tool.poetry]
name = "grpc-try"
version = "0.1.0"
description = ""
authors = ["range <alexadush@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.60.1"
grpcio-tools = "^1.60.1"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
